requirements.txt
flask
requests
datetime
urllib3
streamlit
openai
pytz
python-dotenv
twilio
elasticsearch>=8.0.0,<9.0.0